revert tests changes fix CopyManagerTest.java
