--- src/test/java/cc/blynk/clickhouse/integration/AbstractIntegrationTest.java
@@ -1,34 +1,0 @@
-package cc.blynk.clickhouse.integration;
-
-import cc.blynk.clickhouse.ClickHouseConnection;
-import cc.blynk.clickhouse.ClickHouseDataSource;
-import cc.blynk.clickhouse.settings.ClickHouseProperties;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
-
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.TimeZone;
-
-public class AbstractIntegrationTest {
-
-    protected ClickHouseConnection connection;
-    protected DateFormat dateFormat;
-
-    @BeforeTest
-    public void setUp() throws Exception {
-        ClickHouseProperties properties = new ClickHouseProperties();
-        ClickHouseDataSource dataSource = new ClickHouseDataSource("jdbc:clickhouse://localhost:8123", properties);
-        connection = dataSource.getConnection();
-        connection.createStatement().execute("CREATE DATABASE IF NOT EXISTS test");
-        dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        dateFormat.setTimeZone(TimeZone.getDefault());
-    }
-
-    @AfterTest
-    public void tearDown() throws Exception {
-        connection.createStatement().execute("DROP DATABASE test");
-    }
-
-
-}
